# Specialize BandedMatrix * Diagonal for DenseColumnMajor