Evaluate `B * D` as `B.data .* D.diag'` instead of `B  .* D.diag'`. This helps with vectorization, leading to significant performance improvements.

With arrays
```julia
julia> A = BandedMatrix(0=>Float64[1:3000;]); D = Diagonal(A);

julia> A2 = brand(3000, 3000, 200,200); D2 = Diagonal(A2);
``` 

we obtain
| Operation | Master | PR |
| --- | :-: | :-: |
| `A * D` | 54.393 μs | 2.574 μs |
| `A2 * D2` | 2.252 ms | 1.232 ms |

I've specialized the method for `DenseColumnMajor` layout. It may work with other layouts as well, but this is a conservative first step.